changed definition of consensus slack from abs. int to fraction

+++ include/SalmonDefaults.hpp
@@ -19,7 +19,7 @@ namespace defaults {
 
   // advanced
   constexpr const bool validateMappings{false};
-  constexpr const int32_t consensusSlack{0};
+  constexpr const float consensusSlack{0.0};
   constexpr const double minScoreFraction{0.65};
   constexpr const int16_t matchScore{2};
   constexpr const int16_t mismatchPenalty{-4};
